requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.4
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0 
torchvision==0.16.0